vision: Separate vision loop into separate module
